requirements.txt

flask
rembg
pillow
flask-cors
onnxruntime
